backend/requirements.txt
Flask~=3.1.0
MarkupSafe~=3.0.2
python-dotenv~=1.0.1
flask_sqlalchemy~=3.1.1
psycopg2-binary~=2.9.10
flask-socketio~=5.1.1